Basic card and package namespacing setup
