+++ app/src/androidTest/java/com/ofthemasses/self_management_app/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.self_management_app
+package com.ofthemasses.self_management_app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.self_management_app", appContext.packageName)
+        assertEquals("com.ofthemasses.self_management_app", appContext.packageName)
     }
 }
