+++ app/src/main/java/com/ofthemasses/self_management_app/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.self_management_app.ui.theme
+package com.ofthemasses.self_management_app.ui.theme
 
 import android.app.Activity
 import android.os.Build
