+++ app/src/main/java/com/ofthemasses/self_management_app/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.self_management_app.ui.theme
+package com.ofthemasses.self_management_app.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
